Move properties classes to their corresponding modules
No need to put those in core

COSDK-921
