+++ blik/src/main/java/com/adyen/checkout/blik/internal/ui/view/BlikCodeField.kt
@@ -15,10 +15,10 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.focus.onFocusChanged
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.tooling.preview.PreviewParameter
+import com.adyen.checkout.blik.internal.ui.properties.BlikCodeProperties.BLIK_CODE_MAX_LENGTH
+import com.adyen.checkout.blik.internal.ui.properties.BlikCodeProperties.BLIK_CODE_SEPARATOR
+import com.adyen.checkout.blik.internal.ui.properties.BlikCodeProperties.BLIK_CODE_SEPARATORS
 import com.adyen.checkout.blik.internal.ui.state.BlikIntent
-import com.adyen.checkout.core.common.internal.properties.BlikCodeProperties.BLIK_CODE_MAX_LENGTH
-import com.adyen.checkout.core.common.internal.properties.BlikCodeProperties.BLIK_CODE_SEPARATOR
-import com.adyen.checkout.core.common.internal.properties.BlikCodeProperties.BLIK_CODE_SEPARATORS
 import com.adyen.checkout.core.common.localization.CheckoutLocalizationKey
 import com.adyen.checkout.core.common.localization.internal.helper.resolveString
 import com.adyen.checkout.core.components.internal.ui.state.model.TextInputViewState
