+++ card/src/main/java/com/adyen/checkout/card/internal/ui/view/SocialSecurityNumberOutputTransformation.kt
@@ -10,9 +10,9 @@ package com.adyen.checkout.card.internal.ui.view
 
 import androidx.compose.foundation.text.input.OutputTransformation
 import androidx.compose.foundation.text.input.TextFieldBuffer
-import com.adyen.checkout.core.common.internal.properties.SocialSecurityNumberProperties.CNPJ_SEPARATORS
-import com.adyen.checkout.core.common.internal.properties.SocialSecurityNumberProperties.CPF_SEPARATORS
-import com.adyen.checkout.core.common.internal.properties.SocialSecurityNumberProperties.CPF_VALID_LENGTH
+import com.adyen.checkout.card.internal.ui.properties.SocialSecurityNumberProperties.CNPJ_SEPARATORS
+import com.adyen.checkout.card.internal.ui.properties.SocialSecurityNumberProperties.CPF_SEPARATORS
+import com.adyen.checkout.card.internal.ui.properties.SocialSecurityNumberProperties.CPF_VALID_LENGTH
 import com.adyen.checkout.ui.internal.element.input.SeparatorsTextFieldBufferTransformation
 
 internal class SocialSecurityNumberOutputTransformation : OutputTransformation {
